Replace ScanQrDialog with the KMM ScanQrCodeDialog.
There were significant layout problems with the old QR code reader: the camera
image was often misaligned, the borders were offset, the image could be
squished, and the misalignment was inconsistent. The new QR code scanner looks
much better behaved, though we lose the brackets that used to be there.

Tested by:
- Manual testing of both issuance and identity verification.
- ./gradlew check
- ./gradlew connectedCheck

Signed-off-by: Kevin Deus <kdeus@google.com>