Add Option for Custom OpenID4VP Scheme to Verifier
Added additional protocol - OpenID4VP (Custom URI scheme) -
to the verifier. When that protocol is selected, there is also
a text box which appears next to the dropdown where the user
can input the custom scheme.

Tested manually by registering the wallet app for a random
scheme and using it in the request.

Signed-off-by: Suzanna Jiwani <suzannaj@google.com>