Adds support for direct_post for verification.
Previously we only supported direct_post.jwt. This adds support for
direct_post. The main difference is in OpenID4VPPresentationActivity.kt,
the new getDirectPostAuthorizationResponseBody function (in contrast
to the getDirectPostJwtAuthorizationResponseBody function). Most of the
other changes are just plumbing, or allowing the user to decide between
direct_post and direct_post.jwt.

Tested by:
- Manual testing of both direct_post and direct_post.jwt, using openid4vp://
  and both mdoc and VC.
- ./gradlew check
- ./gradlew connectedCheck

Signed-off-by: Kevin Deus <kdeus@google.com>