B64-encode documentId in routes.
Some document IDs are URLs, so they have slashes in them. That messes with
our route strings, which use slashes to separate parameters. Encoding the
docId fixes a crash when trying to navigate to the Activity Log for a single
event, when using the German or EU PIDs.

Tested by:
- Manual testing of Utopia, German, and EU PIDs.
- ./gradlew check
- ./gradlew connectedCheck

Signed-off-by: Kevin Deus <kdeus@google.com>