Allow null apu or apv in OpenID4VP Response
When a wallet did not provide the apu or apv JWE header
parameters in the response, the verifier would respond with
status code (400). To instead allow the verifier to continue
parsing the response, the apu and apv are checked to see if
null and if so, are not added to the sessionTranscript.

Tested manually against a modified wallet which did
not set apu and apv.

Signed-off-by: Suzanna Jiwani <suzannaj@google.com>