Add multiplatform NFC support.
The foundational parts of this change include low-level routines for
working with APDUs and NDEF messages in the com.android.identity.nfc
package. This is mostly based on code in NfcUtil.kt from
identity-android library. Also add comprehensive unit tests for these
routines.

For reading NFC tags, a new scanNfcTag() function is added. On iOS
this is implemented via NFCTagReader which shows to the user a
bottom-sheet dialog with an application-chosen message. On Android,
similar UI is implemented. Also introduce an abstract `NfcIsoTag`
class to represent the remote NFC tag which can be passed to a
function passed to scanNfcTag() when a tag is in the field.

In the identity-mdoc library, extract the core logic from
VerificationHelper into the mdocReaderNfcHandover() suspend function
which implements the mdoc reader side of NFC engagement according to
18013-5 using a NfcIsoTag. Use this together with scanNfcTag() in a
newly provided scanNfcMdocReader().

In the identity-mdoc library, also extract core logic from
NfcEngagementHelper into MdocNfcEngagementHelper which implements the
mdoc side of NFC engagement according to ISO 18013-5.

Add unit tests for testing the happy path of combining
MdocNfcEngagementHelper with mdocReaderNfcHandover().

In samples/testapp:

- Add screen for scanning a NDEF tag.

- Use scanNfcMdocReader() to implement requesting an mDL using NFC
  engagement.

- Factor out presentation logic to PresentationModel and Presentation
  composable and use this in IsoMdocSharingScreen. Currently only
  supports ISO mdoc proximity presentations but is set up to also
  support other mechanisms including Android Credman, OpenID4VP URI
  scheme, and so on. Think of this as unifying PresentationActivity,
  CredmanPresentationActivity, and OpenID4VPPresentationActivity.
  The intent is to move this to identity-appsupport when it's fully baked.

- On Android, use MdocNfcEngagementHelper to start a newly added
  NfcPresentationActivity which shows the Presentation composable
  and uses PresentationModel. iOS doesn't support host-card emulation
  so this is Android only.

Other changes:

- Rename AndroidInitializer to AndroidContexts and also make this
  singleton object track the currently running Activity. This is
  needed to implement scanNfcTag() so the bottom sheet can be
  shown.

- Change MdocTransportFactory from an object into an interface
  and introduce MdocTransportFactory.Default as an object that
  contains the default implementation for the platform. This
  allows to stub out the factory in tests.

- Fix NfcEngagement.kt in wallet so it works properly with the
  Android NFC disambiguation dialog.

- Bump versions of some dependencies.

Test: Unit tests
Test: Manually tested on Android and iOS
Signed-off-by: David Zeuthen <zeuthen@google.com>