testapp: Fix key material, remove personal build details.
The keys used in the testapp doesn't work with iOS CryptoKit's ASN1
parser (which is a bit finicky) so update them with keys that do
work. The long-term fix for this is to write our own PEM parser
(easily done now that we have a ASN1 library) but for now this fix
will do it.

Also remove references to personal build details.

Test: Manually tested testapp on iOS
Signed-off-by: David Zeuthen <zeuthen@google.com>