Appverifier: Fix NFC engagement problem on first run after install. (#857)
This change addresses an NFC engagement issue that could occur on the first run of AppVerifier after installation. The root cause was an incorrect handling of the Android activity lifecycle events that occur when the permission-requesting activity returns to the main application activity or fragment. The fix ensures that the VerificationHelper instance is initialized during the ON_RESUME lifecycle event, rather than ON_START, to avoid race conditions and ensure proper setup.

Tested manually. Tap with the phone running Wallet.

Signed-off-by: koukarine <koukarine@google.com>