SecureArea: Add mechanism to automatically show authentication dialogs.
Introduce KeyUnlockInteractive which if passed to sign() and
keyAgreement() on SecureArea will show the appropriate authentication
prompts and generate the proper KeyUnlockData-derived objects. Add
support for this in all four SecureArea implementations (Software,
Android Keystore, iOS Secure Enclave, Cloud Secure Area) and port
samples/testapp to use this both in the SecureArea specific screens and
also in PresentmentModel and Presentment composable.

Also make KeyUnlockInteractive the default for SecureArea operations
since it's a vastly simpler model for application programmers,
compared to juggling KeyUnlockData objects and performing user
authentication.

Introduce a PassphrasePromptModel mechanism for the core library to
display passphrase prompt dialogs in the UI layer. This is required for
the Software and Cloud Secure Area implementations of SecureArea when
KeyUnlockInteractive is used.

Add PassphrasePromptBottomSheet composable (based on existing code in
the wallet module) as well as a screen in samples/testapp that demoes
it. Connect this implementation to the PassphrasePromptModel mechanism
via the PassphrasePromptProvider composable.

Start using platform-specific Secure Areas for test documents in
samples/testapp and use two domains so each document has a credential
with user-authentication and one without. Add new "Skip user
authentication" setting in samples/testapp and use this setting to
pick the right domain in TestAppPresentmentSource.

Because of the new KeyUnlockInteractive, we now get the user
authentication part for free in PresentmentModel / Presentment
regardless of the Secure Area in use. In fact, a third party could
create a SecureArea implementation (it's an open class and intended
for this) and as long as they handle UI/UX interactions correctly
things will just work without any changes.

In samples/testapp's screen for Android Keystore and Cloud Secure
Areas, remove the dialogs for displaying certificate chains. Use the
new Certificate Viewer screen instead.

Test: Tested on Android and iOS and all four SecureArea impl
Signed-off-by: David Zeuthen <zeuthen@google.com>