samples/testapp: More fixes for dealing with the world now being async. (#861)
Move all dependencies used by TestApp in the `App` class and make it a
singleton with asynchronous construction.

Use Storage in TestAppSettingsModel to make settings persistent and
introduce a new SettingsScreen and move all settings there. Make this
accessible through a gear icon in the app's top bar.

Startup performance is quite important, especially when the app is
invoked from cold, for example as part of a Credman request or NFC
tap.

We really need initialization of the document database, key material
and certificates used, trust lists, and other items to be in the
handful of milliseconds so add machinery to measure where the time
goes. Use this to learn that generating keys and certificates or even
parsing PEM at startup is prohibitively expensive. Fix this by
serializing these objects in [Storage] so it's only expensive on first
run and practically free for future runs. Ditto, only provision the
documents on the first run of the application.

Make sure the app actually work in cases a Credman or NFC presentment
happen and the app is not currently running. Also make sure it's performant.

For example for the NFC engagament path, defer `App` initialization
until the NFC engagement is actually complete. This is quite important
because timing is very sensitive because the holder's device is only
going to be in the NFC field for so long.

For the Credman case, ensure we can handle the request without first
having re-registered all our documents with Credman which is a costly
affair. This implies storing the Credman registration ID -> Document
mapping in persistent storage. It's tempting to use applicationData on
Document for this but this is going to go away soon so use a separate
database table instead. In order to do this, have `DocumentStore`
expose the `Storage` it's using so we can piggyback and store this new
database table in the same backing file.

Also fix `MdocDataElement.renderValue()` so it understands the ISO/IEC
23220-2 encoding of `birth_date` as used in PhotoID doctype. This
takes care of a few noisy warnings which starting appear when we added
support for PhotoID.

Test: Manually tested.
Test: New unit tests for `MdocDataElement.renderValue()`.

Signed-off-by: David Zeuthen <zeuthen@google.com>