CloudSecureArea: Port to Kotlin multiplatform.
This uses the DeviceCheck API added previously to ensure only the
wallet app authorized to use the Cloud Secure Area can use it. Also
change DeviceCheck to use the term "challenge" instead of "clientId"
as the former more accurate reflects the purpose of the parameter.

Improve the Cloud Secure Area attestation format and add support for
pretty-printing it in CertificateViewer and check that the local keys
are created with the right authentication parameters (e.g. biometric
only, passcode + biometrics).

Move the client-side code and tests in main `identity` library and
delete `identity-android-csa` entirely. Unit tests are still Android
only because it relies on the server-side which is JVM-only right now.

Test: Tested on Android and iOS.
Signed-off-by: David Zeuthen <zeuthen@google.com>