Add support for 18013-7 Annex C and OpenID4VP over W3C DC API.
This includes two new W3C Digital Credentials protocols.

On the wallet side, support is added to PresentmentModel and on the
verifier side the modifications are to VerifierServlet.

This also includes a new Credman matcher, including source code. This
matcher includes copies of cppbor and cJSON libraries.

For OpenID4VP, a simple implementation of DCQL is currently used. The
dcql-library branch has a much more complete implementation which will
be integrated in the near future.

This includes the addition of JsonWebSignature and JsonWebEncryption
objects to implement these standards. This is currently only
implemented on JVM which is not a problem because the W3C Digital
Credentials API is not available on iOS right now.

In preparation for the Utrecht test events, these implementations
(reader and wallet) have been tested against multiple other readers
and wallets with successful outcomes.

Test: Manually tested against 5+ different wallets and readers.
Signed-off-by: David Zeuthen <zeuthen@google.com>