Generalize requesting UI services and add support for biometric and NFC.
This moves PassphrasePromptModel into UiModel and adds a similar
concept in androidMain for Android-specific features including
BiometricPrompt and the dialog to show when scanNfcTag() is called.

This allows downstreams to precisely control how biometric prompts are
rendered and also the look and feel of the NFC scanning dialog, which
is preferable to hard-coding this in the core library.

Also fix a bug in the default provider - now called UiProvider() - so
it tracks the life-cycle so things work properly when using this
machinery from multiple Activities.

Test: Manually tested.
Signed-off-by: David Zeuthen <zeuthen@google.com>