multipaz-compose: Add support for notifications and screen locks.
This adds NotificationManager which can be used to show notifications
to the end user. This includes rememberNotificationPermissionState()
and PermissionState to check for and request the permission. Also
update rememberBluetoothPermissionState() so both use the same
underlying interface. Also fix iOS implementation to properly update
when the Bluetooth permission is granted.

This also adds ScreenLockState and rememberScreenLockState() to track
whether a screen lock has been set up on the device. This is necessary
for dealing with device-bound credentials requiring user
authentication before they can be presented.

Also add screens to samples/testapp to easily manually test all this.

Test: Manually tested on iOS and Android (with and without screen lock on both).
Test: All unit tests pass.
Signed-off-by: David Zeuthen <zeuthen@google.com>