testapp: Add "Document Store" screen for provisioning of documents. (#879)
This also removes document provisioning at start-up which removes the
requirement that a screen lock is needed for the app to work.

The new "Document Store" screen will show all the documents currently
in the store and provides a way to clear the store and also add test
documents, using various Secure Area implementations.

Also make the Android implementation DigitalCredentials listen for
events on DocumentStore so Credman is properly updated as documents
are added and removed.

Also add card art for Photo ID and PID document types.

Also remove compose dependencies from the core library since it's no
longer needed now that PR #874 is in.

A future PR will add a "Document Info" screen to show details about
each Document including the Credentials currently associated with it.

Test: Manually tested on Android and iOS using Platform (Android Keystore resp. Secure Enclave), Software, and Cloud Secure Area.

Signed-off-by: David Zeuthen <zeuthen@google.com>