Fixed an issue where an NFC scan would cause following QR scans to fail. (#884)
We were getting a UiViewNotAvailableException, because UiModel.unregisterView()
was being called when the NFC screen was removed, but UiModel.registerView()
wasn't being called when the underlying screen came back. We were only handling
the ON_START lifecycle event, not the ON_RESUME one, so a Stop/Resume cycle
would leave the UI provider unregistered.

Tested by:
- Manual testing in testapp, where the bug occurred before.

Signed-off-by: Kevin Deus <kdeus@google.com>