Initial changes to prepare files for the package/directory migration.
The build.gradle.kts files in identity, identity-flow, identity-mdoc, and
identity-sdjwt are all being merged, and that's easiest for the migration tool
if they're all identitical in the first place. This changes them so they're all
identical and will work after the migration.

The other changes are also preparation for the automatic migration.

In themselves, these changes break the build. However, after running the
migration script, these have been tested using all the tests listed in
the PR description:

Tested by:
- (see the list in the PR)

Signed-off-by: Kevin Deus <kdeus@google.com>