Fixes and new features added during test events.
IsoMdocProximityReadingScreen wasn't working when the response was
using Key Agreement for mdoc authentication, fixed this.

In samples/testapp, switch to prefering Key Agreement to Signature if
possible when presenting an ISO mdoc. Also add a setting to prefer
using Signature which is handy at test events when dealing with a
reader which doesn't support Key Agreement (most wallets use
Signature). Add PresentmentSource.preferSignatureToKeyAgreement() to
express this preference.

Add keyBound parameter to VcDocumentType to express whether credentials
are intended to be key-bound. Set this to false for UtopiaMovieTicket.
Add MovieTicket to samples/testapp, including new card art. Use
DocumentAttributeType.Date instead of DocumentAttributeType.DateTime
since it's, well, a date, not a date-time.

Fix PhotoID ISO mdoc doctype since it changed in recent 23220-4 drafts.

Fix a problem with NFC Negotiated Handover where the new NFC
engagement code didn't combine / disambiguate properly resulting in
multiple application/vnd.bluetooth.le.oob NDEF records being
transmitted. Ran into this problem with a wallet using NFC Negotiated
Handover. Add unit tests for this and update existing unit tests with
vectors.

Add requirement to SecureArea that SecureArea.getKeyInfo() must
work even on keys that are invalidated. Fix up AndroidKeystoreSecureArea
to honor this requirement.

Extend EcCurve with values of whether a curve supports Key Agreement
and Signing.

Add require() calls to Crypto.keyAgreement() to catch early problems
when app is trying to do this with keys not using the same curve.

Make which curve to use for ISO/IEC 18013-5:2021 session encryption
configurable in samples/testapp, on the Settings page. Also add
settings to the Document Store screen to control which curves are used
for DeviceKey and Document Signing Key when creating new documents.
Also make it possible to indicate of the keys should support Key
Agreement and Signature. All these settings are helpful when at test
events for seeing if a reader supports it since the standard calls for
the reader to support all curves.

In CredentialViewerScreen, add the a way to view the DS certificate(s)
attached to the COSE_Sign1 for the MSO. Also show Device Key curve,
purposes, and whether the key is currently invalidated.

Test: Carefully tested, during four days of test events.
Test: New unit tests.
Test: ./gradlew check && ./gradlew connectedCheck
Signed-off-by: David Zeuthen <zeuthen@google.com>