TestApp: Minor usability fixes.
When first installing samples/testapp it's not intuitive that you need
to go into "Document Store" and add documents so add a warning card to
get the user to do this.

Once you have credentials, it's tempting to just tap the phone against
an mdoc reader but this won't work unless BLE permissions are
granted. So also show a warning urging the user to grant these
permissions.

Also, a partner was trying this on a Pixel 4 and ran into the problem
that KeyMint on this device does not support AGREE_KEY. So add a
better error message instructing the user to turn off KEY_AGREEMENT if
this is the case.

Also get rid of testapp's InfoCard and WarningCard, use the ones from
multipaz-compose instead.

Test: Manually tested on devices with and without KA support in Keymint.
Test: Manually tested on a fresh app install.
Signed-off-by: David Zeuthen <zeuthen@google.com>