Introduce fully-specified Algorithm and use it everywhere. (#909)
As discussed in Issue #850 there's quite a bit of merit to having
SecureArea use fully-specified algorithms as detailed in
[Fully-Specified Algorithms for JOSE and COSE draft](https://datatracker.ietf.org/doc/draft-ietf-jose-fully-specified-algorithms/)
so do this and convert the entire codebase to use it.

It is tempting to also convert the low-level wrappers in Crypto for
the platform cryptographic library to follow this approach, but it's
less clear that is a good idea. This PR does not attempt to do so.

Specifically this means removing KeyPurpose altogether since this is
now conveyed in the algorithm itself. With this change, key creation
boils down to just passing a single Algorithm value, which has to be
fully specified. Apart from being simpler, it also means that it's a
lot easier to add support for other kinds of cryptographic keys to
SecureArea for example RSA (unlikely we'll do this but nice to have
the option) or PQC.

One downside of this is that keys are now single-purpose, a key can no
longer Sign and do Key Agreement, only one of the two. So apps need to
be aware of this and can use different domains for different sets of
keys. Change samples/testapp to do this which means adding two new
domains for ISO mdoc credentials (MAC w/ user auth, MAC w/o user auth)
and also add an option to specify how many credentials to create per
domain. Also rework DocumentStoreScreen so user can pick algorithms
instead of curves. Improve CredentialViewerScreen so it groups
credentials together in domains. Add dialog shown while documents are
created since this can be slow.

Rework both PresentmentSource and its users (ISO mdoc proximity and
W3C DC presentment) to work with this new constraint that a single
credential can be used either for Signing or Key Agreement, never
both.

Rework warnings on the first screen so they're always shown on top
instead of being part of the scrollable area. Also add an
informational card shown when documents are provisioned to convey that
the user can present via QR/NFC and also W3C DC API.

Also clarify how Cose.coseSign1Sign() handles inserting COSE_LABEL_ALG
and add tests for this.

A new supportedAlgorithms property has been added to SecureArea to
convey which algorithms are supported. These algorithms must all be
fully-specified.

Because the format for how key metadata has changed, bump the schema
version and just delete the old data as an "upgrade". It's test data
which users can easily recreate anyway.

Carefully tested proximity presentment with different session
encryption curves and verify that Key Agreement is used exactly when
the session encryption curve matches the curve of DeviceKey in a
credential used for MACing. Also tested W3C DC presentment with all
four supported protocols (preview, ARF, Annex C, OpenID4VP).

Test: Manually tested, see above.
Test: New unit tests and all unit tests pass.

Signed-off-by: David Zeuthen <zeuthen@google.com>