SecureArea: Add new batchCreateKey() method.
This adds a new batchCreateKey() method to create many keys with the
same algorithm in one shot. The default implementation simply calls
createKey() in sequence but implementations may override this. For
example, a Cloud Secure Area implementation can use this to minimize
the amount of network roundtrips. This also adds the possibility to
return an OpenID4VCI key attestation JWT.

Fixes Issue #851 and fixes Issue #852.

Test: New unit test and all unit tests pass.
Signed-off-by: David Zeuthen <zeuthen@google.com>