Updates to cbor serialization annotation processor.
Implemented schema hash which allows to keep track of potentially incompatible changes in
object's CBOR serialization. When a particular object is considered to be stable a
schemaHash parameter to CborSerializable annotation can be specified. schemaHash is a hash
of CBOR serialization schema for the object, when computed and specified value differ
a compile-time error is generated containing the new hash value. If the change deemed safe
by the developer, the specified hash value should be updated in the PR.

Fixed a bug where default CBOR map key to select correct sealed class subclass was set to
"null" instead of "type".

Fixed a bug where code generation failed for internal classes.

Testing: added unit tests for the new functionality.

Signed-off-by: Peter Sorotokin <sorotokin@gmail.com>