Crypto: change checkSignature() to throw if signature verification fails. (#915)
This is the industry-standard and much safer than relying on having
the caller check a boolean. Introduce SignatureVerificationException
to convey this and derive this from SecurityException to form sealed
hierarchy. This is helpful if want to add other security-related
exceptions in the future.

Also add a negative test to check that checkSignature() indeed throws
the expected exception when trying to verify a signature made with
another key.

Test: All unit tests pass.

Signed-off-by: David Zeuthen <zeuthen@google.com>