ConnectionMethod: Properly convey PSM when disambiguating methods.
Commit e0d0f98 introduced a bug where when using NFC Negotiated
Handover and the mdoc reader is offering mdoc BLE Central Client mode
w/ the PSM set (default configuration for our proximity reader in
samples/testapp), we weren't using the PSM from the engagement to
connect directly.

Concretely this manifested itself in slower BLE connections because it
then meant we had to go through the whole GATT dance which the PSM in
the engagement was designed to avoid entirely.

The root problem is that when disambiguating the single received
ConnectionMethodBle object, the PSM was only set on the instance
created for mdoc BLE Peripheral Server mode. The solution to this
problem is to add a new `role` parameter to the disambiguate() call
and use this to control where to put the PSM.

Write unit tests to carefully test this and also test it in
practice. Also tested NFC Static Handover w/ mdoc BLE Peripheral
Server mode and noted that we don't currently convey the PSM in this
case. This fix is more involved and only applies to samples/testapp so
for now just add a TODO so we can revisit later.

Also for ConnectionMethodBle move PSM and MAC address properties into
the constructor for easier access. And use ByteString for the MAC
address.

Test: New unit tests and all unit tests pass.
Test: Manually tested.
Signed-off-by: David Zeuthen <zeuthen@google.com>