testapp: add second flavor and classify red flavor as "payment" app.
This PR adds a second flavor to testapp where the appId is
org.multipaz.testapp.red (instead of org.multipaz.testapp), its name
is "Multipaz Test App (Red)" instead of "Multipaz Test App" and its
icons are red, not blue. Additionally, the red flavor is categorized
as a "payment app" according to
[the Android HCE documentation](https://developer.android.com/develop/connectivity/nfc/hce#groups-categories).
Specifically this means that "Multipaz Test App (red)" is eligble to
be the default Wallet app on Android.

Also add logging to Test App's MainActivity to log the result of
calling CardEmulation.setPreferredService(). Also move AppTheme from
multipaz-compose to samples/testapp as it doesn't make much sense to
have this in the library.

The benefit of this change is that we can use it to easily test that
versions of Android implements the correct behavior when it comes to
handling NFC tag readers selecting a NFC service which both the
Default Wallet App and the current foreground app has registered for.
Simply install both the blue and red flavors, provision them with
documents, and set the red flavor as the default wallet app.

Test: Manually tested.
Signed-off-by: David Zeuthen <zeuthen@google.com>