Fix wallet namespacing and directory structure
The wallet package name was org.multipaz_credential.wallet. It's been changed to org.multipaz.wallet, and the Android ID has been changed to org.multipaz.androidwallet (in anticipation of using org.multipaz.wallet for a cross-platform wallet).

That, plus this moves the code back to non-KMM directories (main and test, rather than androidMain and jvmTest).

Wallet also has an internal "issuance" directory that needed to be renamed to "provisioning" and nested inside a "wallet" namespace so it didn't conflict with the org.multipaz.issuance namespace.

A few references to the new cbor builder functions needed to be fixed up along the way.

A ./gradlew clean may be necessary after merging this PR.

Tested by:
- Android Studio: `Make Project`
- `./gradlew check && ./gradlew connectedCheck`
- Running the wallet app and issing a Utopia driver's license.

Signed-off-by: Kevin Deus <kdeus@google.com>