Change various SecureArea implementations so that data structures that we put in Storage are serialized (#924)
using `@CborSerializable` data classes with schemaHash specified. This will help us to be conscious about
compatibility going forward.

Actual serialized data in the storage should still be the same (although this was not extensively tested)
with the exception of keys in `SoftwareSecureArea` that did not have passphrase (removed duplication
storing both `publicKey` and `encodedPublicKey`).

Testing: unit tests and testing using TestApp

Signed-off-by: Peter Sorotokin <sorotokin@gmail.com>