testapp: Fix use of NFC APDU categories. (#930)
It's not entirely clear from the
[Android documentation](https://developer.android.com/develop/connectivity/nfc/hce)
but it turns out applications should never use CATEGORY_PAYMENT except for
AIDs related to actual payment, e.g. EMVCo protocols.

So change testapp to always use CATEGORY_OTHER for both both blue and red
flavor. This means that neither flavor will appear in the as options
for "Wallet app" in the "Settings -> Apps -> Default apps" screen in
Android. Instead, simply implement the
[Quick Access Wallet](https://source.android.com/docs/core/connect/quick-access-wallet)
interfaces and now both flavors appear as options for Wallet app in
"Default apps".

Also call CardEmulation.setPreferredService() correctly. This resulted
in NdefService being bound early which uncovered a deadlock. Resolve
this by not using `runBlocking` in NdefService, just use a channel to
push APDUs to a coroutine running on an I/O thread instead.

One nice thing about this PR is that both flavors are functionally
equivalent. The only difference is the appId and the color of the
icons.

With these changes everything now works as expected, that is:

- if both flavors are installed and neither are the default Wallet,
  neither are in the foreground, and the default Wallet hasn't
  registered for the NDEF AID, users will see the NFC disambig dialog
  when tapping against an mdoc reader.

- If one flavor is the default wallet, that flavor will be invoked when
  tapping (no disambig dialog) and no other app registering for the NDEF
  AID is in the foreground.

- If one of the flavors is in the foreground, it will always be invoked
  when tapping, no matter if there's a default Wallet app or not.

See https://www.youtube.com/watch?v=8zEcjAgZvNg for a demonstration of this.

Test: Manually tested.

Signed-off-by: David Zeuthen <zeuthen@google.com>