Prevent R8 from removing BouncyCastle components (#944)
Due to the nature of how BouncyCastle is being used as a Java security provider, R8 strips the BC provider and EC algorithm in the release build because it looks like they're not being used in the project.

Add the necessary ProGuard consumer rules to Multipaz to keep them from being removed.

Signed-off-by: Henry Jalonen <henry.jalonen@reaktor.com>