Inject Ktor engine to Multipaz library (#943)
Instead of using explicit CIO engine in :multipaz, inject the HttpClientEngineFactory into CloudSecureArea and remove the io.ktor:ktor-client-cio dependency from the module. This prevents interop issues with projects that already have another engine configured.

In :samples:testapp, provide CIO as the HttpClientEngineFactory, and in :wallet, use Android engine (as it's already used there).

"Use" MockEngine in CloudSecureAreaTest; The engine is provided by the factory, as the HttpClient object is created, but as it's not being used, it will throw an error should it receive a request.

Test: ./gradlew check && ./gradlew connectedCheck

Signed-off-by: Henry Jalonen <henry.jalonen@reaktor.com>