Allow MdocTransport.waitForMessage() cancellation
While the MdocTransport.waitForMessage() execution is being suspended, waiting for the incoming message channel to send new values, allow CancellationExceptions to pass through instead of treating them as errors and mapping them to a MdocTransportExceptions.

This prevents the transport from being erroneously marked as failed, as there's no reason to do so.

Test: ./gradlew check && ./gradlew connectedCheck
Test: Manually tested both wallet and testapp proximity sharing
Signed-off-by: Henry Jalonen <henry.jalonen@reaktor.com>