Avoid mapping CancellationException in MdocTransports (#957)
Add new protected `Throwable.wrapUnlessCancellationException()` extension method to `MdocTransport` that can be used to map exceptions to `MdocTransportExceptions` with a supplied message, unless it's a `CancellationException`, and replace the existing `MdocTransport` logic to use it instead of blindly mapping every exception.

But, to avoid swallowing the BLE stack errors reported by the `onError()` callback, use the Ktor `unwrapCancellationException()` helper to find them (and wrap them).

Replace `CoroutineScope(currentCoroutineContext())` with `coroutineScope` to avoid the `onJoin()` dance and remove the coroutines-internal `getCancellationException()` call. The cancellation exception will be thrown anyway because `coroutineScope` is suspended while the launched job is running.

Test: ./gradlew check connectedCheck
Test: Manually tested the local library build against our implementation
Test: Not tested with Multipaz apps on real devices, though!

Signed-off-by: Henry Jalonen <henry.jalonen@reaktor.com>