Move DEVELOPMENT_TEAM into separate configuration file.
This new configuration file is in .gitignore so with this change
developers will need to set it only once and won't need to unset it
before committing.

Also include a template to easily create this configuration file and
add new DEVELOPER-ENVIRONMENT.md file to help explain what needs to be
done to set up a developer environment including a shell-snippet for
setting up the configuration file.

Test: Manually tested.
Signed-off-by: David Zeuthen <zeuthen@google.com>