Differences in g3 between the imported version and the latest.
The last import was commit ce6cec763e015a34439af9fd5d641600a9fa5014, from
2023-04-08. This takes all the changes that have been done since that import
in the g3 version and ports them to the latest upstream version.

There were many conflicts, mostly from renamed files or files that had been
converted from Java to Kotlin.

The g3 versions had changed the package namespace, which resulted in a number
of unnecessary merge changes. This reverts the package name changes (and
switches to the new multipaz namespace).

The coseSign1CheckNoDuplicateHeaderParams function that had been added to
multipaz-android-legacy/.../Util.kt needed to be moved to where it was used
in DeviceRequestParser.kt, because that's in the core library, multipaz. It
was also converted to be more Kotlin-esque and use standard functions that
exist in DataItem.

Omitted changes:
- The provisionMsosForExistingCredential function in Utility.java has a TODO
  to move it to the mdocstore module. It would take a little work to get This
  function compiling in Multipaz, so instead I have omitted it in hopes that
  the TODO can be fixed and that function can be moved to its final destination.
- The changes in KeystoreIdentityCredential.java depend on a function
  implemented elsewhere in g3, so they were not upstreamed.

Several changes look like they'd already been upstreamed. Two in particular
should be reviewed, however:
1) The g3 version added a isNegotiatedHandover parameter to toNdefRecord. The
   behavior of this flag looked very similar to the new skipUuids pararemeter,
   so this change was not upstreamed.
2) In NfcEngagementHelper.kt, in the handleHandoverRequest function, the g3
   version had added some logic to select the connection method, rather than
   selecting the first disambiguated option. On review, it looked like the
   new disambiguation function already contains similar logic and would
   result in the same connection method, but this needs another set of eyes.

Test: ./gradlew check
Test: ./gradlew connectedCheck
Test: Manually spot-checked screens in testapp: About, Document store, CSA
Test: Testapp mdl proximity reading over NFC
Test: Testapp mdl proximity reading using a QR code
Test: Testapp multi-device testing Happy Flow Short (a typical number were successful)

Signed-off-by: Kevin Deus <kdeus@google.com>