Camera: Rework Camera API and make it work on iOS.
Change onFrameCaptured() so it doesn't return an ImageBitmap which was
intended to be used for the application to overlay graphics. Instead,
if users want to overlay graphics they can simply arrange their
composition so a `Canvas` overlays the `Camera` composable.

Use platform-native types for the returned frame - on Android this is
an ImageProxy and on iOS it's an UIImage. Note that the captured frame
is different sizes/rotation/etc from the preview. To deal with this,
pass a Matrix in CameraFrame so the application can properly transform
into preview space.

Hook this up to barcode scanning and also in CameraScreen have
use-cases that use the preview (drawing an animation on top) and don't
use the previews (just showing an Image with the captured bitmap).

Test: Manually tested barcode scanning on both Android and iOS.
Signed-off-by: David Zeuthen <zeuthen@google.com>